+++ test/examples/gerberx3/api/_30_multi_file_render_with_pillow_defaults_str.example.py
@@ -1,10 +1,9 @@
-from pygerber.gerber.api import GerberFile, Project
+from pygerber.gerber.api import GerberFile, CompositeView
 
 from pygerber.examples import ExamplesEnum, load_example
 
-gerber_source_code = load_example(ExamplesEnum.UCAMCO_2_11_2)
 
-project = Project(
+project = CompositeView(
     [
         GerberFile.from_str(
             load_example(ExamplesEnum.simple_2layer_F_Cu),