+++ test/examples/gerberx3/api/_70_multi_file_constructor_showcase.project.py
@@ -1,10 +1,8 @@
-from pygerber.gerber.api import GerberFile, Project, FileTypeEnum
-
 from pygerber.examples import ExamplesEnum, load_example
+from pygerber.gerber.api import CompositeView, GerberFile, FileTypeEnum
 
-gerber_source_code = load_example(ExamplesEnum.UCAMCO_2_11_2)
 
-project = Project(
+project = CompositeView(
     [
         GerberFile.from_str(
             load_example(ExamplesEnum.simple_2layer_F_Cu),
@@ -24,5 +22,4 @@ project = Project(
         ),
     ],
 )
-image = project.render_with_pillow(dpmm=40)
-image.get_image().save("output.png")
+print(project)