+++ test/examples/gerberx3/api/_72_multi_file_render_with_pillow_single_file_access.example.py
@@ -1,8 +1,8 @@
 from pygerber.examples import ExamplesEnum, load_example
-from pygerber.gerber.api import Project, GerberFile
+from pygerber.gerber.api import CompositeView, GerberFile
 
 
-project = Project(
+project = CompositeView(
     [
         GerberFile.from_str(load_example(ExamplesEnum.simple_2layer_F_Cu)),
         GerberFile.from_str(load_example(ExamplesEnum.simple_2layer_F_Mask)),
@@ -11,4 +11,4 @@ project = Project(
     ],
 )
 image = project.render_with_pillow()
-image.get_image().save("output.png")
+image.get_sub_images()[0].get_image().save("output.png")